[Fix/#62] user_id 캐스팅 문제 해결 (#63)

+++ backend/src/main/kotlin/com/manage/crm/event/domain/repository/EventRepositoryCustomImpl.kt
@@ -78,7 +78,7 @@ class EventRepositoryCustomImpl(
                 Event(
                     id = it["id"] as Long,
                     name = it["name"] as String,
-                    userId = (it["user_id"] as Int).toLong(),
+                    userId = it["user_id"] as Long,
                     properties = (it["properties"] as String)
                         .let { properties ->
                             objectMapper.readValue(properties, List::class.java).stream()
